data_prep/requirements.txt
dask==2022.12.1
pandas==1.5.2
python-dotenv==0.21.0
